fixed Edit, DeleteButton props on BeerDetailsPage
