+++ BeerTier/client/src/components/BeerCard.js
@@ -1,8 +1,5 @@
-import { useNavigate } from "react-router-dom";
-
 import {
   Box,
-  Button,
   Card,
   CardActionArea,
   CardContent,
@@ -15,31 +12,9 @@ import EditButton from "./EditButton";
 import StyleLink from "./StyleLink";
 import UserProfileLink from "./UserProfileLink";
 
-import { deleteFromApi } from "../modules/resourceManager";
+import { deleteBeer } from "../modules/resourceManager";
 
 export default function BeerCard({ beerObj, userProfile }) {
-  const navigate = useNavigate();
-
-  const handleDeleteBeer = (beerId) => {
-    // make sure we pass an object with the required properties
-    const trimmedBeerObj = {
-      id: beerId,
-      name: beerObj.name,
-      breweryId: beerObj.brewery.id,
-      createDateTime: beerObj.createDateTime,
-      userProfileId: beerObj.userProfile.id,
-    };
-
-    const confirmed = window.confirm(
-      "Are you sure you want to delete this beer?"
-    );
-    if (confirmed) {
-      deleteFromApi("beer", trimmedBeerObj, beerId).then(
-        window.location.reload(true)
-      );
-    }
-  };
-
   return (
     <Card className="beer__card" sx={{ mx: "1rem", maxWidth: "50%" }}>
       <CardActionArea href={`beer/${beerObj.id}`}>
@@ -72,11 +47,8 @@ export default function BeerCard({ beerObj, userProfile }) {
           {(userProfile.id === beerObj.userProfile.id ||
             userProfile.isAdmin) && (
             <>
-              <EditButton controller="beer" editObjId={beerObj.id} />
-              <DeleteButton
-                deleteObjId={beerObj.id}
-                deleteHandler={handleDeleteBeer}
-              />
+              <EditButton controller="beer" objRef={beerObj} />
+              <DeleteButton objRef={beerObj} deleteCallback={deleteBeer} />
             </>
           )}
         </CardContent>
