+++ BeerTier/client/src/components/EditButton.js
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 
 import Button from "@mui/material/Button";
 
-export default function EditButton({ controller, editObjId }) {
+export default function EditButton({ controller, objRef }) {
   const navigate = useNavigate();
   // stopPropagation and preventDefault are here to stop tacking on the Card href to the route
   // otherwise you end up with like 'beer/beer/edit/22'
@@ -15,7 +15,7 @@ export default function EditButton({ controller, editObjId }) {
       onClick={(evt) => {
         evt.stopPropagation();
         evt.preventDefault();
-        navigate(`/${controller}/edit/${editObjId}`);
+        navigate(`/${controller}/edit/${objRef.id}`);
       }}
     >
       Edit
