+++ BeerTier/client/src/modules/resourceManager.js
@@ -86,8 +86,8 @@ export const putObjToApi = (resource, obj, id) => {
   });
 };
 
-export const deleteFromApi = (resource, obj, id) => {
-  const apiUrl = `/api/${resource}/${id}`;
+export const deleteFromApi = (resource, obj) => {
+  const apiUrl = `/api/${resource}/${obj.id}`;
   return getToken().then((token) => {
     return fetch(apiUrl, {
       method: "DELETE",
@@ -108,4 +108,24 @@ export const deleteFromApi = (resource, obj, id) => {
       }
     });
   });
+};
+
+// specific delete handlers here
+// for forming specific request bodies
+export const deleteBeer = (beerObj) => {
+  // make sure we pass an object with the required properties
+  const trimmedBeerObj = {
+    id: beerObj.id,
+    name: beerObj.name,
+    breweryId: beerObj.brewery.id,
+    createDateTime: beerObj.createDateTime,
+    userProfileId: beerObj.userProfile.id,
+  };
+
+  const confirmed = window.confirm(
+    "Are you sure you want to delete this beer?"
+  );
+  if (confirmed) {
+    deleteFromApi("beer", trimmedBeerObj).then(window.location.reload(true));
+  }
 };
